Al eliminar acuerdo no se elimina la beca

+++ app/models/agreement.rb
@@ -46,6 +46,6 @@ class Agreement < ApplicationRecord
   end
 
   def delete_agreeable
-    self.agreeable.destroy
+    self.agreeable.destroy unless self.agreeable_type.eql?'Scholarship'
   end
 end
